add WorkerWrapperBase
